requirements.txt
# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.12.0
tensorflow>=2.8.0

# Data processing
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Model serving and API
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0

# Monitoring and logging
mlflow>=1.24.0
wandb>=0.12.0
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0

# Code quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.950

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
click>=8.0.0
tqdm>=4.64.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

